add tx count and thread id to stats (#20451) each stat reports and resets when slot changes
